Merging revision 2 changes into master branch

+++ code/scripts/higher_order_sims.py
@@ -6,26 +6,28 @@ import os
 from config import config
 import timecorr as tc
 from matplotlib import pyplot as plt
+import seaborn as sns
 
-cond= sys.argv[1]
-r= sys.argv[2]
+sim_function = sys.argv[1]
+r = sys.argv[2] #reps
 
-F = 4 #number of features
-T = 30 #number of timepoints
+F = int(sys.argv[3]) #number of features
+T = int(sys.argv[4]) #number of timepoints
 K = 2 #order
+W = int(sys.argv[5])
+wp = sys.argv[6]
 
-fname = cond + '_' + str(F) + '_' + str(T) + '_' + str(K)
+fname = sim_function
 
-width = 20
 
-results_dir = os.path.join(config['resultsdir'], 'higher_order_sims', cond)
+results_dir = os.path.join(config['resultsdir'], 'higher_order_sims',
+                           sim_function)
 
 try:
     if not os.path.exists(results_dir):
         os.makedirs(results_dir)
 except OSError as err:
    print(err)
-
 
 
 def expanded_vec2mat(v):
@@ -36,7 +38,11 @@ def expanded_vec2mat(v):
   return x
 
 
-laplace = {'name': 'Laplace', 'weights': tc.laplace_weights, 'params': {'scale': width}}
+laplace = {'name': 'Laplace', 'weights': tc.laplace_weights, 'params': {'scale': W}}
+gaussian = {'name': 'Gaussian', 'weights': tc.gaussian_weights, 'params': {'var': W}}
+mexican_hat = {'name': 'Mexican hat', 'weights': tc.mexican_hat_weights, 'params': {'sigma': W}}
+
+weights_paramter = eval(wp)
 
 eye_params = {}
 
@@ -50,17 +56,19 @@ def generate_templates(order=1, **kwargs):
   T = kwargs['T']
   templates = []
   for n in range(order - 1):
+    print(n)
     templates.append(next_template)
 
     expanded_corrmats = tc.vec2mat(next_template)
     K2 = expanded_corrmats.shape[0] ** 2
     next_template = np.zeros([K2, K2, T])
     for t in range(T):
-      x = np.atleast_2d(expanded_corrmats[:, :, t].ravel())
-      next_template[:, :, t] = x * x.T
+      x = expanded_corrmats[:, :, t]
+      next_template[:, :, t] = np.kron(x, x)
     next_template = tc.mat2vec(next_template)
   templates.append(next_template)
   return templates
+
 
 
 def generate_data(templates):
@@ -69,6 +77,7 @@ def generate_data(templates):
     next_corrmats = adjusted_templates[-1]
 
     for n in range(order - 1, 1, -1):
+        print(n)
         corrmats = tc.vec2mat(next_corrmats)
         K = corrmats.shape[0]
         sK = int(np.sqrt(K))
@@ -76,6 +85,7 @@ def generate_data(templates):
 
         draws = np.zeros([sK, sK, T])
         means = tc.vec2mat(templates[n - 2])
+
 
         for t in range(T):
             draws[:, :, t] = np.reshape(np.random.multivariate_normal(means[:, :, t].ravel(), corrmats[:, :, t]),
@@ -102,7 +112,7 @@ if not os.path.exists(save_file):
 
     recovery_performance_all = pd.DataFrame()
 
-    templates = generate_templates(order=K, S=1, T=T, K=F, datagen=cond)
+    templates = generate_templates(order=K, S=1, T=T, K=F, datagen=sim_function)
 
     data, adjusted_templates = generate_data(templates)
 
@@ -112,33 +122,41 @@ if not os.path.exists(save_file):
     recovery_performance = pd.DataFrame(index=np.arange(T), columns=np.arange(1, K+1))
     recovery_performance.index.name = 'time'
     recovery_performance.columns.name = 'order'
+
     next_data = data
     recovered_corrs_raw = []
     recovered_corrs_smooth = []
+
     for k in np.arange(1, K+1):
-      next_recovered_smooth = tc.timecorr(next_data, weights_function=laplace['weights'], weights_params=laplace['params'])
+      next_recovered_smooth = tc.timecorr(next_data, weights_function=weights_paramter['weights'],
+                                    weights_params=weights_paramter['params'])
+
       next_recovered_raw = tc.timecorr(next_data, weights_function=eye_weights, weights_params=eye_params)
       recovered_corrs_smooth.append(next_recovered_smooth)
       F_new = get_f(next_recovered_smooth.shape[1])
       for t in np.arange(T):
         recovery_performance.loc[t, k] = np.corrcoef(templates[k-1][t, F_new:], next_recovered_smooth[t, F_new:])[0, 1]
 
+
+
       next_data = expanded_vec2mat(next_recovered_raw)
 
-    # recovery_performance.columns = [str(x + 1) for x in np.arange(K)]
-    #
-    # recovery_performance['iteration'] = int(r)
-
-    #recovery_performance_all = recovery_performance_all.append(recovery_performance)
+    recovery_performance.columns = [str(x + 1) for x in np.arange(K)]
+    recovery_performance['iteration'] = int(r)
+    recovery_performance_all = recovery_performance_all.append(recovery_performance)
 
     print(recovery_performance)
+    plt.clf()
+    plt.plot(recovery_performance['1'])
+    plt.plot(recovery_performance['2'])
+    plt.show()
 
-    recovery_performance.to_csv(save_file + '.csv')
+    plt.savefig(results_dir + '2')
 
-# if not os.path.isfile(save_file + '.csv'):
-#     recovery_performance.to_csv(save_file + '.csv')
-# else:
-#     append_iter = pd.read_csv(save_file + '.csv', index_col=0)
-#     append_iter = append_iter.append(recovery_performance)
-#     append_iter.to_csv(save_file + '.csv')
 
+    if not os.path.isfile(save_file + '.csv'):
+        recovery_performance.to_csv(save_file + '.csv')
+    else:
+        append_iter = pd.read_csv(save_file + '.csv', index_col=0)
+        append_iter = append_iter.append(recovery_performance)
+        append_iter.to_csv(save_file + '.csv')
